backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.21.0
pydantic-settings==2.10.1
httpx==0.28.1
python-dotenv==1.1.1
PyJWT==2.10.1
aiogram==3.20.0.post0
APScheduler==3.11.0
